requirements.txt
pytest==8.4.1
docker==7.1.0
requests==2.32.3
pyyaml==6.0.2